requirements.txt
alembic==1.14.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==33.1.0
Flask==3.1.0
Flask-Migrate==4.0.7
flask-openapi3==4.0.3
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.8
MarkupSafe==3.0.2
openai==1.58.1
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
pytest==8.3.4
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
